[04] - implemented ObjectConfigurator.kt
