+++ src/main/kotlin/com/vlopatka/reflection/objectConfigurator/ObjectConfigurator.kt
@@ -1,5 +1,5 @@
 package com.vlopatka.reflection.objectConfigurator
 
 interface ObjectConfigurator {
-    fun <T> configure(t: Class<T>)
+    fun  configure(obj: Any)
 }
